fix dynamic height on new tabs
